Moved generic util functions from server to lib + fixed invalid imports
